Added settings dialog to manage settings w/ pyqtconfig, my fixed version
Attention, it's currently not working with the version that is on the
pyqtconfig github and it sure doesn't work with the version on PyPi
since that one does not yet support PyQt6. I opened a PR for the
bug fix that is required in order to use the settings dialog.